refactor: Comment CRUD 관련 함수들 분리
