+++ src/entities/comment/ui/CommentIcons.tsx
@@ -1,20 +1,19 @@
 import { Edit2, ThumbsUp, Trash2 } from "lucide-react";
 import { Button } from "../../../shared/ui";
 import { Comment } from "../model/types";
+import { useComment } from "../lib/useComment";
 
 interface CommentIconsProps {
   comment: Comment;
-  onLikeComment: (commentId: number, postId: number) => void;
   onEditComment: (comment: Comment) => void;
-  onDeleteComment: (commentId: number, postId: number) => void;
 }
 
 export const CommentIcons: React.FC<CommentIconsProps> = ({
   comment,
-  onLikeComment,
   onEditComment,
-  onDeleteComment,
 }) => {
+  const { handleDeleteComments, handleLikeComments } = useComment();
+
   return (
     <div className="flex items-center space-x-1">
       <Button
@@ -22,7 +21,7 @@ export const CommentIcons: React.FC<CommentIconsProps> = ({
         size="sm"
         onClick={() => {
           if (comment?.id && comment.postId)
-            onLikeComment(comment?.id, comment.postId);
+            handleLikeComments(comment?.id, comment.postId);
         }}
       >
         <ThumbsUp className="w-3 h-3" />
@@ -40,7 +39,7 @@ export const CommentIcons: React.FC<CommentIconsProps> = ({
         size="sm"
         onClick={() => {
           if (comment?.id && comment.postId)
-            onDeleteComment(comment.id, comment.postId);
+            handleDeleteComments(comment.id, comment.postId);
         }}
       >
         <Trash2 className="w-3 h-3" />
